requirements.txt
moondream==0.0.1
streamlit>=1.31.0
Pillow>=7.1.0,<11.0.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.12.2
requests==2.32.3
tqdm>=4.66.0